requirements.txt
Flask
numpy
pandas
scikit-learn
matplotlib
opencv-python-headless
tensorflow-gpu
Exit